更改TitleBar高度为double 右键菜单加入关闭组件功能

+++ DesktopWidgets3/ViewModels/Pages/Widget/FrameShellViewModel.cs
@@ -11,7 +11,7 @@ public partial class FrameShellViewModel : ObservableRecipient
     }
 
     [ObservableProperty]
-    private int _widgetTitleBarHeight = 0;
+    private double _widgetTitleBarHeight = 0;
 
     public FrameShellViewModel(IWidgetNavigationService widgetNavigationService)
     {
